// types.ts
export interface Media {
  id: string;
  type: 'image' | 'video';
  url: string; // Data URL
  name: string;
  observation?: string; 
}
// ...
export interface AgendaItem {
  id: string;
  userId: string; // Vínculo obrigatório com o usuário para privacidade
  title: string;
  description: string;
  createdAt: string; // ISO String
  dueDate: string;   // ISO String (Data/Hora do Lembrete)
  status: 'Pending' | 'Done';
  notified: boolean; // Controle para o alerta sonoro não repetir
  lastEmailDate?: string; // Data do último e-mail de alerta enviado
}

export interface AgendaIssue {
  id: string;
  userId: string;
  date: string;
  clientName: string;
  description: string;
  media: Media[];
  status: 'Pending' | 'Resolved';
  createdAt: string;
}
// ...
export enum UserRole {
  MASTER = 'Master',
  OPERATIONAL = 'Operacional'
}

export interface User {
  id: string;
  username: string;
  password: string;
  role: UserRole;
}